Extracted aliases code from interpreter.cpp
