+++ CMakeLists.txt
@@ -416,7 +416,8 @@ set(SOURCES
 		src/engine/ui/cmd_god/do_spellstat.cpp
 		src/gameplay/statistics/spell_usage.cpp
 		src/engine/ui/cmd_god/do_loadstat.cpp
-		src/engine/ui/cmd_god/do_sanitize.cpp)
+		src/engine/ui/cmd_god/do_sanitize.cpp
+		src/engine/ui/cmd/do_alias.cpp)
 
 set(HEADERS
 		src/engine/ui/cmd/do_flee.h
@@ -820,7 +821,8 @@ set(HEADERS
 		src/engine/ui/cmd_god/do_spellstat.h
 		src/gameplay/statistics/spell_usage.h
 		src/engine/ui/cmd_god/do_loadstat.h
-		src/engine/ui/cmd_god/do_sanitize.h)
+		src/engine/ui/cmd_god/do_sanitize.h
+		src/engine/ui/cmd/do_alias.h)
 
 # Build types
 set(CMAKE_CONFIGURATION_TYPES Release Debug Test FastTest)
