+++ src/engine/ui/alias.h
@@ -9,9 +9,25 @@
 #ifndef BYLINS_SRC_ENGINE_UI_ALIAS_H_
 #define BYLINS_SRC_ENGINE_UI_ALIAS_H_
 
+struct alias_data {
+  char *alias;
+  char *replacement;
+  int type;
+  struct alias_data *next;
+};
+
+inline constexpr int kAliasSimple{0};
+inline constexpr int kAliasComplex{1};
+inline constexpr char kAliasSepChar{'+'};
+inline constexpr char kAliasVarChar{'='};
+inline constexpr char kAliasGlobChar{'*'};
+
 class CharData;
-void write_aliases(CharData *ch);
-void read_aliases(CharData *ch);
+void WriteAliases(CharData *ch);
+void ReadAliases(CharData *ch);
+void FreeAlias(struct alias_data *a);
+int PerformAlias(DescriptorData *d, char *orig);
+struct alias_data *FindAlias(struct alias_data *alias_list, char *str);
 
 #endif //BYLINS_SRC_ENGINE_UI_ALIAS_H_
